src/requirements.txt
python-telegram-bot==13.1
urlextract==1.0.0
youtube_dl
